Added authorization to all contacControllers

+++ controllers/contactControllers.js
@@ -3,15 +3,15 @@ const Contact = require('../models/contactModel');
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access Public
+//@access Private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
     console.log('200 OK');
 });
 //@desc Get single contact
 //@route GET /api/contacts/:id
-//@access Public
+//@access Private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -23,7 +23,7 @@ const getContact = asyncHandler(async (req, res) => {
 });
 //@desc Create new contacts
 //@route POST /api/contacts
-//@access Public
+//@access Private
 const createContact = asyncHandler(async (req, res) => {
     console.log('The request body is: ', req.body);
     const { name, email, phone } = req.body;
@@ -35,18 +35,24 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     res.status(201).json(contact);
     console.log('201 OK');
 });
 //@desc Update a single
 //@route PUT /api/contacts
-//@access Public
+//@access Private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User doesn't have persmission to update this contact");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -65,6 +71,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User doesn't have persmission to update this contact");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
